requirements.txt
scipy==1.0.1
python-igraph==0.7.1.post6
